# truck-service-host/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Truck Service API Documentation",
        "description": "Interactive API documentation for the Truck Service application. Developed by Mohammed Sleem.",
        "contact": {
            "email": "dev.mohammed@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://bull-station.com",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/trucks/{truck}": {
            "get": {
                "tags": [
                    "Public Data"
                ],
                "summary": "Get details of a single truck",
                "operationId": "getTruckDetails",
                "parameters": [
                    {
                        "name": "truck",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Truck details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TruckResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found or not active.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not found or not active."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Truck Management"
                ],
                "summary": "Update an existing truck",
                "description": "To update a truck with files (images/video), you must send a POST request and include `_method=PATCH` in the form-data. This is a limitation of how browsers handle file uploads.",
                "operationId": "updateTruck",
                "parameters": [
                    {
                        "name": "truck",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "enum": [
                                            "PATCH"
                                        ],
                                        "example": "PATCH"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1600.50"
                                    },
                                    "images[]": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Truck updated and awaiting re-approval.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Truck updated and awaiting re-approval."
                                        },
                                        "truck_id": {
                                            "type": "integer",
                                            "example": 123
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Truck not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Truck not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Truck Management"
                ],
                "summary": "Delete a truck",
                "description": "Requires authentication. Allows a truck owner to delete their own truck.",
                "operationId": "deleteTruck",
                "parameters": [
                    {
                        "name": "truck",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Truck has been successfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Truck has been successfully deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Truck not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Truck not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User Model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "account_type": {
                        "type": "string",
                        "enum": [
                            "client",
                            "truck_owner"
                        ]
                    }
                },
                "type": "object"
            },
            "TruckResource": {
                "title": "Truck Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "active",
                            "inactive"
                        ]
                    },
                    "model": {
                        "type": "string"
                    },
                    "year_of_manufacture": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price_per_day": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "price_per_hour": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "work_hours": {
                        "type": "string",
                        "example": "08:00:00 - 18:00:00"
                    },
                    "pickup_location": {
                        "type": "string"
                    },
                    "delivery_price": {
                        "type": "string",
                        "format": "decimal"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "type": "string"
                    },
                    "sub_category": {
                        "type": "string"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "url"
                        }
                    },
                    "video": {
                        "type": "string",
                        "format": "url"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}